+++ core/admin/AJAX.php
@@ -2,7 +2,7 @@
 
 namespace Dev4Press\Plugin\DemoPress\Admin;
 
-use Dev4Press\v49\Core\Quick\WPR;
+use Dev4Press\v50\Core\Quick\WPR;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;