+++ core/data/text/LorIpsumNet.php
@@ -2,8 +2,8 @@
 
 namespace Dev4Press\Plugin\DemoPress\Data\Text;
 
-use Dev4Press\v49\Core\Options\Element as EL;
-use Dev4Press\v49\Core\Options\Type;
+use Dev4Press\v50\Core\Options\Element as EL;
+use Dev4Press\v50\Core\Options\Type;
 use WP_Error;
 
 if ( ! defined( 'ABSPATH' ) ) {