+++ core/generator/bbPress.php
@@ -2,9 +2,9 @@
 
 namespace Dev4Press\Plugin\DemoPress\Generator;
 
-use Dev4Press\v49\Core\Options\Element as EL;
-use Dev4Press\v49\Core\Options\Type;
-use function Dev4Press\v49\Functions\sanitize_ids_list;
+use Dev4Press\v50\Core\Options\Element as EL;
+use Dev4Press\v50\Core\Options\Type;
+use function Dev4Press\v50\Functions\sanitize_ids_list;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;