+++ core/library/Randomizer.php
@@ -3,7 +3,7 @@
 namespace Dev4Press\Plugin\DemoPress\Library;
 
 use Dev4Press\Plugin\DemoPress\Base\Library;
-use Dev4Press\v49\Generator\Text\Randomizer as BaseRandomizer;
+use Dev4Press\v50\Generator\Text\Randomizer as BaseRandomizer;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;